fix: display site menu in safari
The site menu wasn't being rendered correctly in the safari browser due
to some issue with `position: fixed` being applied to the element.

This patch changes the `position: fixed` to a `position: relative` and
fiddles with the layout and margins a bit to get the menu to go to the
correct position.